requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
h11==0.14.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.0
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.19.2
